add a payload file for storing request payload structs

+++ src/server.cc
@@ -92,7 +92,7 @@ int main(int argc, char** argv) {
 
                 rdma_request rdma_req(num_cols);
 
-                for (int i = 0; i < num_cols; i++) {
+                for (int64_t i = 0; i < num_cols; i++) {
                     std::shared_ptr<arrow::Array> col_arr = batch->column(i);
                     arrow::Type::type type = col_arr->type_id();
                     int64_t num_rows = col_arr->length();
